remove passing consent defaults into init, fix circular deps
